Split out model
